frontend
This project is built with React and React-DOM useful for building and rendering the user interface.
Axios is used for making HTTP requests to interact with backend APIs.
Vite -A fast build tool and development server is used for an efficient development process.
TailwindCSS is used for rapidly styling the application.